project-null/Enchanters/backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Redis
redis[hiredis]==5.2.1

# Celery
celery[redis]==5.4.0

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# JWT / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# HTTP client
httpx==0.28.1

# PDF generation
reportlab==4.2.5
pdfrw==0.4

# AWS S3
boto3==1.35.86

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Misc
python-dotenv==1.0.1
